// types/prisma.ts
export enum UserRole {
  HOSPITAL_ADMIN = 'HOSPITAL_ADMIN',
  SURGEON = 'SURGEON',
  NURSE = 'NURSE',
  FULFILLMENT_AGENT = 'FULFILLMENT_AGENT',
  FULFILLMENT = 'FULFILLMENT',
  SUBMITTER = 'SUBMITTER',
  ADMIN = 'ADMIN',
}

export enum OrderStatus {
  PENDING = 'PENDING',
  IN_PROGRESS = 'IN_PROGRESS',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED',
}